Move fixtures back to conftest
